move circuit value logic to its own file
