+++ src/lib/int.ts
@@ -1,10 +1,5 @@
 import { Field, Bool } from './core.js';
-import {
-  AnyConstructor,
-  CircuitValue,
-  Struct,
-  prop,
-} from './provable-types/struct.js';
+import { AnyConstructor, Struct } from './provable-types/struct.js';
 import { Types } from '../bindings/mina-transaction/types.js';
 import { HashInput } from './hash.js';
 import { Provable } from './provable.js';
@@ -14,6 +9,7 @@ import { addMod32 } from './gadgets/arithmetic.js';
 import type { Gadgets } from './gadgets/gadgets.js';
 import { withMessage } from './field.js';
 import { FieldVar } from './provable-core/fieldvar.js';
+import { CircuitValue, prop } from './provable-types/circuit-value.js';
 
 // external API
 export { UInt8, UInt32, UInt64, Int64, Sign };
